move stations
