+++ routes/__init__.py
@@ -1,1 +1,1 @@
-from . import general
+from . import general, stations
